Arduino IR Relay Motor Controller
I designed this code to operate three AC voltage motors through IR sensors. I have been trying to add a Serial Connection to an ESP8266 to control it through a WiFi connection, but I haven't gotten it to work yet. I'm very new to this. Feel free to branch code as much as you want.